refactor BaseViewModel
