+++ mobile/src/main/java/org/horizontal/tella/mobile/mvvm/settings/CollectServersViewModel.kt
@@ -2,28 +2,22 @@ package org.horizontal.tella.mobile.mvvm.settings
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
 import com.google.firebase.crashlytics.FirebaseCrashlytics
 import dagger.hilt.android.lifecycle.HiltViewModel
 import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.schedulers.Schedulers
 import org.horizontal.tella.mobile.R
 import org.horizontal.tella.mobile.data.database.DataSource
 import org.horizontal.tella.mobile.data.database.KeyDataSource
 import org.horizontal.tella.mobile.data.openrosa.OpenRosaService
 import org.horizontal.tella.mobile.domain.entity.collect.CollectServer
+import org.horizontal.tella.mobile.mvvm.base.BaseSettingsViewModel
 import javax.inject.Inject
 
 @HiltViewModel
 class CollectServersViewModel @Inject constructor(
     private val keyDataSource: KeyDataSource
-) : ViewModel() {
-
-    private val disposables = CompositeDisposable()
-
-    private val _loading = MutableLiveData<Boolean>()
-    val loading: LiveData<Boolean> get() = _loading
+) : BaseSettingsViewModel() {
 
     private val _listCollectServers = MutableLiveData<List<CollectServer>>()
     val listCollectServers: LiveData<List<CollectServer>> get() = _listCollectServers
@@ -37,8 +31,6 @@ class CollectServersViewModel @Inject constructor(
     private val _removedServer = MutableLiveData<CollectServer>()
     val removedServer: LiveData<CollectServer> get() = _removedServer
 
-    private val _error = MutableLiveData<Int>()
-    val error: LiveData<Int> get() = _error
 
     fun getCollectServers() {
         disposables.add(
@@ -52,7 +44,6 @@ class CollectServersViewModel @Inject constructor(
                     { servers -> _listCollectServers.postValue(servers) },
                     { throwable ->
                         FirebaseCrashlytics.getInstance().recordException(throwable)
-                       // _error.postValue(throwable)
                     }
                 )
         )
@@ -69,8 +60,7 @@ class CollectServersViewModel @Inject constructor(
                 .subscribe(
                     { createdServer -> _createdServer.postValue(createdServer) },
                     { throwable ->
-                        FirebaseCrashlytics.getInstance().recordException(throwable)
-                        _error.postValue(R.string.settings_docu_toast_fail_create_server)
+                        handleError(throwable, R.string.settings_docu_toast_fail_create_server)
                     }
                 )
         )
@@ -90,8 +80,7 @@ class CollectServersViewModel @Inject constructor(
                         _updatedServer.postValue(updatedServer)
                     },
                     { throwable ->
-                        FirebaseCrashlytics.getInstance().recordException(throwable)
-                        _error.postValue(R.string.settings_docu_toast_fail_update_server)
+                        handleError(throwable, R.string.settings_docu_toast_fail_update_server)
                     }
                 )
         )
@@ -115,8 +104,7 @@ class CollectServersViewModel @Inject constructor(
                         _removedServer.postValue(server)
                     },
                     { throwable ->
-                        FirebaseCrashlytics.getInstance().recordException(throwable)
-                        _error.postValue(R.string.settings_docu_toast_fail_delete_server)
+                        handleError(throwable, R.string.settings_docu_toast_fail_delete_server)
                     }
                 )
         )
