+++ mobile/src/main/java/org/horizontal/tella/mobile/mvvm/settings/DropBoxServersViewModel.kt
@@ -2,34 +2,25 @@ package org.horizontal.tella.mobile.mvvm.settings
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
 import com.google.firebase.crashlytics.FirebaseCrashlytics
 import dagger.hilt.android.lifecycle.HiltViewModel
 import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.schedulers.Schedulers
 import org.horizontal.tella.mobile.R
 import org.horizontal.tella.mobile.data.database.DropBoxDataSource
 import org.horizontal.tella.mobile.data.database.KeyDataSource
 import org.horizontal.tella.mobile.data.openrosa.OpenRosaService
 import org.horizontal.tella.mobile.domain.entity.dropbox.DropBoxServer
+import org.horizontal.tella.mobile.mvvm.base.BaseSettingsViewModel
 import javax.inject.Inject
 
 @HiltViewModel
 class DropBoxServersViewModel @Inject constructor(
     private val keyDataSource: KeyDataSource,
-) : ViewModel() {
-    private val disposables = CompositeDisposable()
-
-    // LiveData to observe changes
-    private val _loading = MutableLiveData<Boolean>()
-    val loading: LiveData<Boolean> get() = _loading
+) : BaseSettingsViewModel() {
 
     private val _listDropBoxServers = MutableLiveData<List<DropBoxServer>>()
     val listDropBoxServers: LiveData<List<DropBoxServer>> get() = _listDropBoxServers
-
-    private val _error = MutableLiveData<Int>()
-    val error: LiveData<Int> get() = _error
 
     private val _serverCreated = MutableLiveData<DropBoxServer>()
     val serverCreated: LiveData<DropBoxServer> get() = _serverCreated
@@ -48,8 +39,7 @@ class DropBoxServersViewModel @Inject constructor(
                 .subscribe(
                     { servers -> _listDropBoxServers.postValue(servers) },
                     { throwable ->
-                        FirebaseCrashlytics.getInstance().recordException(throwable)
-                     //   _error.postValue(throwable)
+                        handleError(throwable)
                     }
                 )
         )
@@ -70,8 +60,7 @@ class DropBoxServersViewModel @Inject constructor(
                 .subscribe(
                     { createdServer -> _serverCreated.postValue(createdServer) },
                     { throwable ->
-                        FirebaseCrashlytics.getInstance().recordException(throwable)
-                        _error.postValue(R.string.settings_docu_toast_fail_create_server)
+                        handleError(throwable, R.string.settings_docu_toast_fail_create_server)
                     }
                 )
         )
@@ -95,8 +84,7 @@ class DropBoxServersViewModel @Inject constructor(
                         _serverRemoved.postValue(server)
                     },
                     { throwable ->
-                        FirebaseCrashlytics.getInstance().recordException(throwable)
-                        _error.postValue(R.string.settings_docu_toast_fail_delete_server)
+                        handleError(throwable, R.string.settings_docu_toast_fail_delete_server)
                     }
                 )
         )
