+++ mobile/src/main/java/org/horizontal/tella/mobile/mvvm/settings/UwaziServersViewModel.kt
@@ -2,27 +2,21 @@ package org.horizontal.tella.mobile.mvvm.settings
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import com.google.firebase.crashlytics.FirebaseCrashlytics
 import dagger.hilt.android.lifecycle.HiltViewModel
 import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.schedulers.Schedulers
 import org.horizontal.tella.mobile.R
 import org.horizontal.tella.mobile.data.database.KeyDataSource
 import org.horizontal.tella.mobile.data.database.UwaziDataSource
 import org.horizontal.tella.mobile.data.openrosa.OpenRosaService
 import org.horizontal.tella.mobile.domain.entity.UWaziUploadServer
+import org.horizontal.tella.mobile.mvvm.base.BaseSettingsViewModel
 import javax.inject.Inject
 
 @HiltViewModel
 class UwaziServersViewModel @Inject constructor(
     private val keyDataSource: KeyDataSource
-) : ViewModel() {
-    private val disposables = CompositeDisposable()
-
-    private val _loading = MutableLiveData<Boolean>()
-    val loading: LiveData<Boolean> get() = _loading
+) : BaseSettingsViewModel() {
 
     private val _listUwaziServers = MutableLiveData<List<UWaziUploadServer>>()
     val listUwaziServers: LiveData<List<UWaziUploadServer>> get() = _listUwaziServers
@@ -36,9 +30,6 @@ class UwaziServersViewModel @Inject constructor(
     private val _removedServer = MutableLiveData<UWaziUploadServer>()
     val removedServer: LiveData<UWaziUploadServer> get() = _removedServer
 
-    private val _error = MutableLiveData<Int>()
-    val error: LiveData<Int> get() = _error
-
     fun getUwaziServers() {
         disposables.add(
             keyDataSource.uwaziDataSource
@@ -50,8 +41,8 @@ class UwaziServersViewModel @Inject constructor(
                 .subscribe(
                     { servers -> _listUwaziServers.postValue(servers) },
                     { throwable ->
-                        FirebaseCrashlytics.getInstance().recordException(throwable)
-                        //  _error.postValue(throwable)
+                        handleError(throwable)
+
                     }
                 )
         )
@@ -68,8 +59,7 @@ class UwaziServersViewModel @Inject constructor(
                 .subscribe(
                     { createdServer -> _createdServer.postValue(createdServer) },
                     { throwable ->
-                        FirebaseCrashlytics.getInstance().recordException(throwable)
-                        _error.postValue(R.string.settings_docu_toast_fail_create_server)
+                        handleError(throwable, R.string.settings_docu_toast_fail_create_server)
                     }
                 )
         )
@@ -86,8 +76,7 @@ class UwaziServersViewModel @Inject constructor(
                 .subscribe(
                     { updatedServer -> _updatedServer.postValue(updatedServer) },
                     { throwable ->
-                        FirebaseCrashlytics.getInstance().recordException(throwable)
-                        _error.postValue(R.string.settings_docu_toast_fail_update_server)
+                        handleError(throwable, R.string.settings_docu_toast_fail_update_server)
                     }
                 )
         )
@@ -111,15 +100,11 @@ class UwaziServersViewModel @Inject constructor(
                         _removedServer.postValue(server)
                     },
                     { throwable ->
-                        FirebaseCrashlytics.getInstance().recordException(throwable)
-                        _error.postValue(R.string.settings_docu_toast_fail_delete_server)
+                        handleError(throwable, R.string.settings_docu_toast_fail_delete_server)
                     }
                 )
         )
     }
 
-    override fun onCleared() {
-        disposables.dispose()
-        super.onCleared()
-    }
+
 }
